refactor dashboard into reusable section components
